refactor: change file structure, put route paths into enum
