+++ src/app-styles/fonts.scss
@@ -4,7 +4,7 @@
   font-style: normal;
   font-weight: 700;
   font-display: swap;
-  src: url(../../../assets/fonts/gilroy/Gilroy-Bold.woff2) format('woff2');
+  src: url(../assets/fonts/gilroy/Gilroy-Bold.woff2) format('woff2');
 }
 
 /* gilroy semibold */
@@ -13,7 +13,7 @@
   font-style: normal;
   font-weight: 600;
   font-display: swap;
-  src: url(../../../assets/fonts/gilroy/Gilroy-Semibold.woff2) format('woff2');
+  src: url(../assets/fonts/gilroy/Gilroy-Semibold.woff2) format('woff2');
 }
 
 /* gilroy regular */
@@ -22,7 +22,7 @@
   font-style: normal;
   font-weight: 400;
   font-display: swap;
-  src: url(../../../assets/fonts/gilroy/Gilroy-Regular.woff2) format('woff2');
+  src: url(../assets/fonts/gilroy/Gilroy-Regular.woff2) format('woff2');
 }
 
 /* gilroy light */
@@ -31,7 +31,7 @@
   font-style: normal;
   font-weight: 300;
   font-display: swap;
-  src: url(../../../assets/fonts/gilroy/Gilroy-Light.woff2) format('woff2');
+  src: url(../assets/fonts/gilroy/Gilroy-Light.woff2) format('woff2');
 }
 
 /* gilroy ultralight */
@@ -40,5 +40,5 @@
   font-style: normal;
   font-weight: 200;
   font-display: swap;
-  src: url(../../../assets/fonts/gilroy/Gilroy-UltraLight.woff2) format('woff2');
+  src: url(../assets/fonts/gilroy/Gilroy-UltraLight.woff2) format('woff2');
 }
