+++ src/components/LogoLink/LogoLink.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import styles from './styles.module.scss';
-import logo from '../../../assets/images/header/logo.png';
+import logo from '../../assets/images/header/logo.png';
 import { NavLink } from 'react-router-dom';
 
 const LogoLink = () => {
