+++ src/components/WeatherWidget/index.ts
