+++ src/components/WeatherWidget/styles.module.scss
