+++ src/enums/routes.ts
@@ -1,0 +1,4 @@
+export const enum Routes {
+  Home = '/',
+  NotFound = '*',
+}
