+++ src/modules/PrimarySection/styles.module.scss
@@ -3,7 +3,7 @@
 }
 
 .backgroundContainer {
-  background-image: url('../../../assets/images/header/main-pic.jpg');
+  background-image: url('../../assets/images/header/main-pic.jpg');
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
